# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: OpenSeaFeed API
  version: "1.0"
  description: |
    The open, community-owned AIS network. All HTTP endpoints live on
    `https://api.openseafeed.com`; the live WebSocket stream lives on
    `wss://stream.openseafeed.com/v1/stream`; raw NMEA contribution goes to
    `ingest.openseafeed.com`.

    ## Keys and tiers

    Keys are created on the [dashboard](https://api.openseafeed.com/) after
    signing in with GitHub or an email sign-in link.

    | kind | prefix | purpose |
    |---|---|---|
    | live | `osf_live_...` | consuming: stream, snapshot, history |
    | feed | `osf_feed_...` | contributing an upstream feed |
    | station | `osf_stn_...` | contributing your own receiver (via station registration) |

    **Anonymous access is welcome** at the free tier: no key needed for the
    stream (bounding boxes up to 30000 square degrees of area), single-vessel state,
    recent history (last 48 h, downsampled) and the delayed fleet snapshot.
    Accounts that actively contribute a feed or receiver get the
    `contributor` tier: unlimited stream area, the freshest snapshot, and
    unclamped history.

    ## Getting data: the WebSocket stream

    Connect to `wss://stream.openseafeed.com/v1/stream` and send one JSON
    subscription within 3 seconds (aisstream.io-compatible; resend at any
    time to swap the subscription):

    ```json
    {
      "APIKey": "osf_live_...",
      "BoundingBoxes": [[[54, -5], [72, 32]]],
...
    }
    ```

...

...

...

...

...

...

...
servers:
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: query
      name: key
      description: An `osf_live_...` key. Omit entirely for anonymous free-tier access.
    sessionCookie:
      type: apiKey
      in: cookie
      name: osf_session
      description: Signed session cookie set by the sign-in flows.
  schemas:
    Vessel:
      type: object
      description: Current state of one vessel, from the snapshotter.
      properties:
        mmsi: { type: integer, example: 257970000 }
        lat: { type: number, example: 59.488 }
        lon: { type: number, example: 21.397 }
        sog: { type: number, description: Speed over ground, knots }
        cog: { type: number, description: Course over ground, degrees }
        hdg: { type: integer, description: True heading, degrees }
        nav: { type: integer, description: AIS navigational status code }
        name: { type: string, example: XANTHIA }
        type: { type: integer, description: ITU ship-type code, example: 80 }
        imo: { type: integer, example: 9246152 }
        callsign: { type: string }
        dest: { type: string, description: Declared destination }
        draught: { type: number, description: Metres }
        len: { type: integer, description: Length overall, metres }
        beam: { type: integer, description: Beam, metres }
        eta:
          type: object
          description: Raw AIS ETA block
          properties:
            Month: { type: integer }
            Day: { type: integer }
            Hour: { type: integer }
            Minute: { type: integer }
        ts: { type: integer, description: Unix milliseconds of last update }
...
paths:
  /v1/vessels/{mmsi}:
    get:
      tags: [vessels]
      summary: Current state of one vessel
      description: Anonymous-friendly. Warm across restarts (state is rebuilt from history).
      security: [{}, { apiKey: [] }]
      parameters:
        - { name: mmsi, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Current state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Vessel" }
        "404": { description: Vessel not seen in the last 24 h }
  /v1/snapshot:
    get:
      tags: [vessels]
      summary: Full-fleet snapshot
      description: |
        Every vessel's current state as one gzip-compressed JSON document,
        regenerated continuously. Anonymous/free tier receives a snapshot at
        least 10 minutes old; contributors get the newest.
      security: [{}, { apiKey: [] }]
      responses:
        "200":
          description: Gzip-compressed JSON array of Vessel objects
          headers:
            x-osf-generated-at: { schema: { type: integer }, description: Unix ms }
            x-osf-vessels: { schema: { type: integer } }
...